requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
bcrypt
python-jose
python-multipart
python-dotenv
boto3
fastapi-limiter
redis
websockets